requirements.txt
fastapi
uvicorn
streamlit
requests
google-genai
python-dotenv
chromadb
pypdf
google-cloud-aiplatform
